Desaclopado metodo para validacao de componente
